requirements.txt
openai
python-binance
pandas
numpy
pyyaml
requests
websockets
python-telegram-bot
aiohttp
tqdm
rich
